feat: create date format helper
